chore(prophet-finder): rename component files to kebab-case and update imports per AGENTS.md
