+++ apps/torus-prophet-finder/src/app/_components/cards/cards-header.tsx
@@ -1,4 +1,4 @@
-import type { EntityMode } from "~/app/_components/cards/EntityModeToggle";
+import type { EntityMode } from "~/app/_components/cards/entity-mode-toggle";
 
 interface Props {
   mode: EntityMode;
