+++ apps/torus-prophet-finder/src/app/_components/cards/empty-state.tsx
