+++ apps/torus-prophet-finder/src/app/_components/cards/entity-mode-toggle.tsx
