+++ apps/torus-prophet-finder/src/app/_components/cards/sample-prophets.ts
