+++ apps/torus-prophet-finder/src/app/_components/cards/search-input.tsx
