+++ apps/torus-prophet-finder/src/app/_components/cards/stable-height.tsx
