+++ apps/torus-prophet-finder/src/app/_components/hero/scroll-indicator.tsx
