+++ apps/torus-prophet-finder/src/app/_components/profile-card.tsx
