+++ apps/torus-prophet-finder/src/app/_components/tickers/fracture-art.tsx
